Fix non-functional firmware update button on K24 SC
On K24 SOM-based SC boards using the new QSPI layout, the
image recovery offset address in the Image Selector App was
not updated. As a result, the firmware update push button
appeared non-functional because the recovery image could not
be located. This patch updates the image recovery offset to
match the new QSPI layout, restoring correct firmware update
functionality via the push button.

Signed-off-by: Dasari Sharath Kumar <sharath.kumar.dasari@amd.com>